Fix tcp metadata not saved
